+++ SharpSqlBuilder/Blocks/SelectColumnBlock.cs
@@ -10,7 +10,7 @@ namespace SharpSqlBuilder.Blocks
     ///     SqlColumn AS PropertyName block
     ///     <example>column AS PropertyName</example>
     /// </summary>
-    public class SelectColumnBlock : SqlBuilderEntity
+    public class SelectColumnBlock : SelectColumnBlockBase
     {
         public readonly Operand TableColumn;
         public readonly OutputValueEntity Value;
@@ -26,16 +26,30 @@ namespace SharpSqlBuilder.Blocks
             TableColumn = operand;
             Value = value;
         }
+        public SelectColumnBlock(Operand operand, string @as)
+        {
+            TableColumn = operand;
+            Value = new OutputValueEntity(@as);
+        }
+
+        public SelectColumnBlock(Operand operand)
+        {
+            TableColumn = operand;
+        }
 
         public override bool Present(SqlOptions sqlOptions) =>
-            TableColumn.Present(sqlOptions) && Value.Present(sqlOptions);
+            TableColumn.Present(sqlOptions);
 
         public override string BuildSql(SqlOptions sqlOptions)
         {
             var tableColumnName = TableColumn.BuildSql(sqlOptions, FlowOptions.Construct(this));
+            if (Value?.Present(sqlOptions)!=true)
+                return $"{tableColumnName}";
+
             var property = Value.BuildSql(sqlOptions, FlowOptions.Construct(this));
             var command = sqlOptions.Command("AS");
             return $"{tableColumnName} {command} {property}";
+
         }
     }
 }