+++ SharpSqlBuilder/Blocks/SelectColumnBlockBase.cs
@@ -1,0 +1,10 @@
+﻿namespace SharpSqlBuilder.Blocks
+{
+    /// <summary>
+    ///     SqlColumn AS PropertyName block
+    ///     <example>column AS PropertyName</example>
+    /// </summary>
+    public abstract class SelectColumnBlockBase : SqlBuilderEntity
+    {
+    }
+}