+++ SharpSqlBuilder/Operands/AggregatableEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace SharpSqlBuilder.Operands
+{
+    public abstract class AggregatableEntity : SqlBuilderEntity
+    {
+    }
+}