+++ SharpSqlBuilder/Operands/AggregateFunctions/CountAggregate.cs
@@ -1,0 +1,15 @@
+﻿namespace SharpSqlBuilder.Operands
+{
+
+    /// <summary>
+    /// <example>COUNT(...)</example>
+    /// </summary>
+    public class CountAggregate : AggregateOperand
+    {
+        public CountAggregate(AggregatableEntity aggregatableEntity) : base(aggregatableEntity)
+        {
+        }
+
+        protected override string Command { get; } = "COUNT";
+    }
+}