+++ SharpSqlBuilder/Operands/AggregateFunctions/MinAggregate.cs
@@ -1,0 +1,14 @@
+﻿namespace SharpSqlBuilder.Operands
+{
+    /// <summary>
+    /// <example>MIN(...)</example>
+    /// </summary>
+    public class MinAggregate : AggregateOperand
+    {
+        public MinAggregate(AggregatableEntity aggregatableEntity) : base(aggregatableEntity)
+        {
+        }
+
+        protected override string Command { get; } = "MIN";
+    }
+}