+++ SharpSqlBuilder/Operands/StarEntity.cs
@@ -1,0 +1,10 @@
+﻿namespace SharpSqlBuilder.Operands
+{
+    public class StarEntity : SqlBuilderEntity
+    {
+        public override bool Present(SqlOptions sqlOptions) => true;
+
+        public override string BuildSql(SqlOptions sqlOptions) => "*";
+        public static readonly StarEntity Instance = new StarEntity();
+    }
+}